Backfill js tests
