+++ src/js/components/githubUsers/assignTaskRole.tsx
@@ -101,6 +101,7 @@ const AssignTaskRoleModal = ({
     setFindText('');
   };
   const handleSave = () => {
+    /* istanbul ignore else */
     if (selection) {
       setUser(selection, shouldAlertAssignee);
     }
